Move utils functions only used for testing to a new class HadoopFSTestUtils
