+++ hudi-hadoop-common/src/main/java/org/apache/hudi/hadoop/fs/HadoopFSUtils.java
@@ -19,9 +19,6 @@
 
 package org.apache.hudi.hadoop.fs;
 
-import org.apache.hadoop.conf.Configurable;
-import org.apache.hadoop.mapred.FileInputFormat;
-import org.apache.hadoop.mapred.JobConf;
 import org.apache.hudi.avro.model.HoodieFSPermission;
 import org.apache.hudi.avro.model.HoodieFileStatus;
 import org.apache.hudi.avro.model.HoodiePath;
@@ -152,20 +149,12 @@ public class HadoopFSUtils {
     return providedPath;
   }
 
-  public static Configuration convertToHadoopConf(StorageConfiguration<?> storageConf) {
-    return storageConf.unwrapAs(Configuration.class);
-  }
-
   /**
    * @param path {@link StoragePath} instance.
    * @return the Hadoop {@link Path} instance after conversion.
    */
   public static Path convertToHadoopPath(StoragePath path) {
     return new Path(path.toUri());
-  }
-
-  public static JobConf convertToJobConf(StorageConfiguration<?> storageConf) {
-    return new JobConf(storageConf.unwrapAs(Configuration.class));
   }
 
   /**
@@ -213,10 +202,6 @@ public class HadoopFSUtils {
         pathInfo.getBlockSize(),
         pathInfo.getModificationTime(),
         convertToHadoopPath(pathInfo.getPath()));
-  }
-
-  public static void setConfForConfigurableInputFormat(FileInputFormat inputFormat, Configuration conf) {
-    ((Configurable) inputFormat).setConf(conf);
   }
 
   /**
