+++ hudi-hadoop-common/src/main/java/org/apache/hudi/storage/hadoop/HadoopStorageConfiguration.java
@@ -38,10 +38,6 @@ public class HadoopStorageConfiguration extends StorageConfiguration<Configurati
 
   private transient Configuration configuration;
 
-  public HadoopStorageConfiguration() {
-    this(true);
-  }
-
   public HadoopStorageConfiguration(Boolean loadDefaults) {
     this(new Configuration(loadDefaults));
   }
